% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_votes.R
\name{animate_votes}
\alias{animate_votes}
\title{Animate vote probabilities for a single row}
\usage{
animate_votes(sim, row = 1)
}
\arguments{
\item{sim}{simulate_forest() output}

\item{row}{row index}
}
\description{
Animate vote probabilities for a single row
}
