% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_accuracy_growth.R
\name{plot_accuracy_growth}
\alias{plot_accuracy_growth}
\title{Plot performance over trees (OOB and optional test set)}
\usage{
plot_accuracy_growth(sim)
}
\arguments{
\item{sim}{Output from simulate_forest()}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ggplot object
}\if{html}{\out{</div>}}
}
\description{
Plot performance over trees (OOB and optional test set)
}
