% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_acc_groth_multi.R
\name{plot_accuracy_growth_multi}
\alias{plot_accuracy_growth_multi}
\title{Plot performance growth across multiple forest runs}
\usage{
plot_accuracy_growth_multi(sim_multi)
}
\arguments{
\item{sim_multi}{Output from \code{simulate_forest_multi()}}
}
\value{
A ggplot object
}
\description{
Produces a line plot showing either accuracy or MSE across trees for multiple
random forest simulations. Each \verb{(run, dataset)} pair is a separate line.
OOB and Test are shown using different linetypes, but share the same color per run.
}
