% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_snapshots.R
\name{tree_snapshots}
\alias{tree_snapshots}
\title{Extract a tree and convert it to a data.frame}
\usage{
tree_snapshots(sim, tree = 1)
}
\arguments{
\item{sim}{simulate_forest() output}

\item{tree}{integer; which tree}
}
\description{
Extract a tree and convert it to a data.frame
}
