% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_accuracy_growth.R
\name{plot_accuracy_growth}
\alias{plot_accuracy_growth}
\title{Plot performance over trees (OOB-Accuracy or MSE and optional test set)}
\usage{
plot_accuracy_growth(sim)
}
\arguments{
\item{sim}{Output from simulate_forest()}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ggplot object
}\if{html}{\out{</div>}}
}
\description{
Produces a line plot showing either accuracy or MSE across trees for a
random forest simulation.
}
\examples{
# Sample 70\% of rows for training
sim <- simulate_forest_multi(
 data = ChickWeight, 
 target = "weight", 
 ntree = 150,
 mtry = 2,
 test_fraction = 0.3,
 test_frac_seed = 42
)
# run the funtion to plot the result
plot_accuracy_growth(sim)
# plot another dataset
sim2 <- simulate_forest(
  data = head(ggplot2::diamonds, n = 1000),
  target = "cut",
  ntree = 100,
  mtry = 2,
  test_fraction = 0.5,
  )
plot_accuracy_growth(sim2)
}
