% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_acc_groth_multi.R
\name{plot_accuracy_growth_multi}
\alias{plot_accuracy_growth_multi}
\title{Plot performance growth across multiple forest runs}
\usage{
plot_accuracy_growth_multi(sim_multi)
}
\arguments{
\item{sim_multi}{Output from \code{simulate_forest_multi()}}
}
\value{
A ggplot object
}
\description{
Produces a line plot showing either accuracy or MSE across trees for multiple
random forest simulations. Each \verb{(run, dataset)} pair is a separate line.
OOB and Test are shown using different linetypes, but share the same color
per run.
}
\examples{
# Sample 70\% of rows for training
sim <- simulate_forest_multi(
 data = ChickWeight, 
 target = "weight", 
 ntree = 150,
 mtry = 2,
 n_runs = 4,
 seeds = c(42, 43, 44, 45),
 test_fraction = 0.3,
)
plot_accuracy_growth_multi(sim)
# plot another dataset
sim2 <- simulate_forest_multi(
  data = head(ggplot2::diamonds, n = 1000),
  n_runs = 4,
  target = "cut",
  ntree = 100,
  mtry = 2,
  test_fraction = 0.5,
  )
plot_accuracy_growth_multi(sim2)
}
