% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_forest.R
\name{simulate_forest}
\alias{simulate_forest}
\title{Incrementally train a random forest and record performance}
\usage{
simulate_forest(
  data,
  target,
  ntree = 100,
  mtry = NULL,
  test_fraction = NULL,
  test_frac_seed = NULL,
  test = NULL
)
}
\arguments{
\item{data}{Training data.frame with predictors and target}

\item{target}{Name of target column (string)}

\item{ntree}{Number of trees to grow}

\item{mtry}{Number of variables randomly sampled at each split}

\item{test_fraction}{Optional, float between 1 and 0, that indicated the
fraction of the datasset to be used as a test}

\item{test_frac_seed}{Seed to make data set split reproducible}

\item{test}{Optional test set (must include target column)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{  tibble with tree-level performance (OOB + test) and model
}\if{html}{\out{</div>}}
}
\description{
Incrementally train a random forest and record performance
}
\examples{
# Train the random forest model using the simulate_forest function
result <- simulate_forest(
  data = iris,
  target = "Species",
  ntree = 100,
  mtry = 2,
  test_fraction = 0.3,
  test_frac_seed = 42
)
}
