% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_forest_multi.R
\name{simulate_forest_multi}
\alias{simulate_forest_multi}
\title{Run multiple random forest simulations in parallel}
\usage{
simulate_forest_multi(
  data,
  target,
  ntree = 100,
  n_runs = 4,
  seeds = NULL,
  mtry = NULL,
  test_fraction = NULL,
  test = NULL
)
}
\arguments{
\item{data}{Training data.frame with predictors and target}

\item{target}{Name of target column (string)}

\item{ntree}{Number of trees to grow}

\item{n_runs}{Number of simulations to run (default: 4)}

\item{seeds}{Optional vector of random seeds for each thread; will
be generated if NULL, this does also influence the
selection of testing datata when using (test_fraction)}

\item{mtry}{Number of variables randomly sampled at each split}

\item{test_fraction}{Optional, float between 1 and 0, that indicated the
fraction of the datasset to be used as a test
Excludes use of (test) parameter!}

\item{test}{Optional test set (must include target column), for
testing the forest with seperate testing data.
Excludes use of (test_fraction) parameter!}
}
\value{
A data frame combining all runs, with an added \code{run} column
}
\description{
This is a wrapper around \code{simulate_forest()} that runs it multiple times in
parallel.
}
\examples{
# Train 4 random forest model using the simulate_forest function
result1 <- simulate_forest_multi(
  data = iris,
  target = "Species",
  ntree = 100,
  mtry = 2,
  test_fraction = 0.3,
  n_runs = 4,
  seeds = c(42, 43, 44, 45)
)
# Train the random forest model using the diamonds dataset
result2 <- simulate_forest_multi(
 data = head(ggplot2::diamonds, n = 1000),
 n_runs = 4,
 target = "cut",
 ntree = 100,
 test_fraction = 0.5,
 )
 plot_accuracy_growth_multi(result)
}
